As we all are well versed with the saying “Health is wealth”, a healthy nation is a wealthy nation.  The quality of healthcare system is considered as one of the crucial factors that define a country’s  state and is also highly recognized as a great contributor to the country’s economy. Statistical  modelling helps us to determine where our research needs to be focused and realize the factors  that impact the healthcare quality and expenses. Also, there are many challenges to be  considered while maintaining the health quality of a country. One such challenges is the health  expenditure. The main objective of our research and analysis is to understand how to minimize  the health expenses while improving or maintaining the same health quality.  Collection of data and data source play a paramount role in the analysis. Since analysis relies on  data, the data considered must be legitimate, else the entire analysis would go in vain. To prevent  such an event in this report, we have built our own dataset by taking the data from reliable  resources such as OECD and WHO after analyzing the literature.
As we all are well versed with the saying “Health is wealth”, a healthy nation is a wealthy nation. 
The quality of healthcare system is considered as one of the crucial factors that define a country’s 
state and is also highly recognized as a great contributor to the country’s economy. Statistical 
modelling helps us to determine where our research needs to be focused and realize the factors 
that impact the healthcare quality and expenses. Also, there are many challenges to be 
considered while maintaining the health quality of a country. One such challenges is the health 
expenditure. The main objective of our research and analysis is to understand how to minimize 
the health expenses while improving or maintaining the same health quality. 
Collection of data and data source play a paramount role in the analysis. Since analysis relies on 
data, the data considered must be legitimate, else the entire analysis would go in vain. To prevent 
such an event in this report, we have built our own dataset by taking the data from reliable 
resources such as OECD and WHO after analyzing the literature. 
After collecting the data and analyzing it, we aim to understand the factors impacting the life 
expectancy and health expenditure. We also aim to provide recommendations to reduce the 
health expenses and improving the health quality of the country. For this, we performed an 
exploratory analysis, built hypothesis and regression models to analyze the data. Two 
independent models have been built to understand and analyze the factors that are affecting 
healthcare quality and expenses. The statistical data modelling involved building up of several 
models and to find out the unbiased model that generalizes and provide feasible 
recommendations that could reduce the health expenses and improve the health quality. 
The key insights from our models is that the life expectancy of a country can be improved by 
increasing the health resources such as hospitals, medical graduates. The country’s healthcare 
expenses can have the maximum benefit when we bring more % of population under public 
insurance. 
Problem Definition & Significance 
We would like to present our findings and suggestions to the policy makers and top health 
officials of the OECD countries. Although, all countries under consideration are developed 
countries, these countries failed to formulate good strategic health policies, these resulted in 
high expenditure over the years with minimal improvement in Quality. 
Our model provides recommendations to these countries of the factors they need to consider 
while formulating policies that improve Health care quality and keeping the expenses low. This is 
an interesting problem that might explain the peculiar case of USA which spends around 10000$ 
per person as compared to other OECD countries that spend an average of $3000, while Health 
quality measures such as life expectancy is not significantly different